(PUP-2173) Move UTF-8 integration tests to config app unit tests
